// src/types/types.tsx
export interface Students {
  _id: string;
  fullName: string;
  email: string;
  phone: string;
  interest: string;
  haveALaptop: boolean;
  createdAt: string;
  updatedAt: string;
}

export interface Partners {
  _id: string;
  name: string;
  email: string;
  phone: string;
  howWouldYouLikeToPartner: string;
  createdAt: string;
  updatedAt: string;
}